Get rid of sign-compare warnings
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@1274 599ab11a-0111-0410-8a50-d7a5b0953570

+++ orange/src/orange.c
@@ -111,7 +111,7 @@ static bool copy_file(const char* input_filename, const char* output_filename)
 
   while (bytes_left)
   {
-    long bytes_to_copy = MIN(bytes_left, BUFFER_SIZE);
+    unsigned bytes_to_copy = MIN(bytes_left, BUFFER_SIZE);
 
     if (bytes_to_copy != fread(buffer, 1, bytes_to_copy, input_file))
       goto exit;
